test(application-generic): Refactor mock cache service into separate file
