+++ packages/application-generic/src/services/cache/index.ts
@@ -2,3 +2,4 @@ export * from './invalidate-cache.service';
 export * from './cache.service';
 export * from './interceptors';
 export * from './key-builders';
+export { MockCacheService } from './cache-service.mock';
